+++ packages/pytest-simcore/src/pytest_simcore/helpers/httpx_calls_capture_errors.py
@@ -1,0 +1,15 @@
+class CaptureProcessingError(Exception):
+    # base for all the exceptions in this submodule
+    pass
+
+
+class VerbNotInPathError(CaptureProcessingError):
+    pass
+
+
+class PathNotInOpenApiSpecError(CaptureProcessingError):
+    pass
+
+
+class OpenApiSpecError(CaptureProcessingError):
+    pass
