+++ packages/pytest-simcore/src/pytest_simcore/helpers/httpx_client_base_dev.py
@@ -8,11 +8,8 @@ from httpx._types import URLTypes
 from jsonschema import ValidationError
 from pydantic import parse_file_as
 
-from .httpx_calls_capture import (
-    CaptureProcessingError,
-    HttpApiCallCaptureModel,
-    get_captured,
-)
+from .httpx_calls_capture_errors import CaptureProcessingError
+from .httpx_calls_capture_model import HttpApiCallCaptureModel, get_captured
 
 _logger = logging.getLogger(__name__)
