+++ services/api-server/tests/unit/api_studies/test_api_routes_studies.py
@@ -12,7 +12,7 @@ import pytest
 from faker import Faker
 from fastapi import status
 from pydantic import parse_file_as, parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
 from respx import MockRouter
 from simcore_service_api_server.models.schemas.errors import ErrorGet
 from simcore_service_api_server.models.schemas.studies import Study, StudyID, StudyPort
