+++ services/api-server/tests/unit/api_studies/test_api_routes_studies_jobs.py
@@ -15,7 +15,7 @@ import respx
 from faker import Faker
 from fastapi import status
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
 from respx import MockRouter
 from simcore_service_api_server._meta import API_VTAG
 from simcore_service_api_server.models.schemas.jobs import Job, JobOutputs
