+++ services/api-server/tests/unit/captures/test__mocks_captures.py
@@ -15,13 +15,11 @@ import jsonref
 import pytest
 import respx
 from pydantic import parse_file_as
-from pytest_simcore.helpers.httpx_calls_capture import (
+from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_openapi import _determine_path
+from pytest_simcore.helpers.httpx_calls_capture_parameters import (
     CapturedParameter,
-    HttpApiCallCaptureModel,
     PathDescription,
-)
-from simcore_service_api_server.utils.http_calls_capture_processing import (
-    _determine_path,
 )
 
 try:
