+++ services/api-server/tests/unit/conftest.py
@@ -38,7 +38,7 @@ from moto.server import ThreadedMotoServer
 from packaging.version import Version
 from pydantic import HttpUrl, parse_obj_as
 from pytest_mock.plugin import MockerFixture
-from pytest_simcore.helpers.httpx_calls_capture import (
+from pytest_simcore.helpers.httpx_calls_capture_model import (
     HttpApiCallCaptureModel,
     PathDescription,
 )
