+++ services/api-server/tests/unit/test_api_solver_jobs.py
@@ -20,7 +20,7 @@ from models_library.api_schemas_webserver.resource_usage import PricingUnitGet
 from models_library.api_schemas_webserver.wallets import WalletGetWithAvailableCredits
 from models_library.generics import Envelope
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
 from simcore_service_api_server._meta import API_VTAG
 from simcore_service_api_server.models.schemas.jobs import Job, JobStatus
 from simcore_service_api_server.models.schemas.solvers import Solver
