+++ services/api-server/tests/unit/test_api_wallets.py
@@ -9,7 +9,7 @@ from fastapi import status
 from httpx import AsyncClient
 from models_library.api_schemas_webserver.wallets import WalletGetWithAvailableCredits
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
 from simcore_service_api_server._meta import API_VTAG
 from unit.conftest import SideEffectCallback
